AI rework: prepare for the future changes and pick AI at runtime (#499)
