+++ fheroes2.vcxproj
@@ -186,11 +186,13 @@
     <ClCompile Include="src\fheroes2\agg\sprite.cpp" />
     <ClCompile Include="src\fheroes2\agg\til.cpp" />
     <ClCompile Include="src\fheroes2\agg\xmi.cpp" />
-    <ClCompile Include="src\fheroes2\ai\ai_action.cpp" />
-    <ClCompile Include="src\fheroes2\ai\simple\ai_battle.cpp" />
-    <ClCompile Include="src\fheroes2\ai\simple\ai_castle.cpp" />
-    <ClCompile Include="src\fheroes2\ai\simple\ai_heroes.cpp" />
-    <ClCompile Include="src\fheroes2\ai\simple\ai_kingdom.cpp" />
+    <ClCompile Include="src\fheroes2\ai\ai_hero_action.cpp" />
+    <ClCompile Include="src\fheroes2\ai\ai_base.cpp" />
+    <ClCompile Include="src\fheroes2\ai\normal\ai_normal.cpp" />
+    <ClCompile Include="src\fheroes2\ai\simple\ai_simple_battle.cpp" />
+    <ClCompile Include="src\fheroes2\ai\simple\ai_simple_castle.cpp" />
+    <ClCompile Include="src\fheroes2\ai\simple\ai_simple_heroes.cpp" />
+    <ClCompile Include="src\fheroes2\ai\simple\ai_simple_kingdom.cpp" />
     <ClCompile Include="src\fheroes2\ai\simple\ai_simple.cpp" />
     <ClCompile Include="src\fheroes2\army\army.cpp" />
     <ClCompile Include="src\fheroes2\army\army_bar.cpp" />
@@ -383,6 +385,8 @@
     <ClInclude Include="src\fheroes2\agg\til.h" />
     <ClInclude Include="src\fheroes2\agg\xmi.h" />
     <ClInclude Include="src\fheroes2\ai\ai.h" />
+    <ClInclude Include="src\fheroes2\ai\empty\ai_empty.h" />
+    <ClInclude Include="src\fheroes2\ai\normal\ai_normal.h" />
     <ClInclude Include="src\fheroes2\ai\simple\ai_simple.h" />
     <ClInclude Include="src\fheroes2\army\army.h" />
     <ClInclude Include="src\fheroes2\army\army_bar.h" />
