+++ src/fheroes2/ai/ai.h
@@ -38,6 +38,13 @@ namespace Battle
 
 namespace AI
 {
+    enum AI_TYPE
+    {
+        EMPTY,
+        SIMPLE,
+        NORMAL
+    };
+
     enum modes_t
     {
         HEROES_MOVED = 0x08000000,
@@ -47,38 +54,43 @@ namespace AI
         HEROES_STUPID = 0x80000000
     };
 
-    void Init( void );
+    class Base
+    {
+    public:
+        virtual void KingdomTurn( Kingdom & );
+        virtual void CastleTurn( Castle & );
+        virtual void BattleTurn( Battle::Arena &, const Battle::Unit &, Battle::Actions & );
+        virtual void HeroesTurn( Heroes & );
 
-    void KingdomTurn( Kingdom & );
-    void BattleTurn( Battle::Arena &, const Battle::Unit &, Battle::Actions & );
-    bool BattleMagicTurn( Battle::Arena &, const Battle::Unit &, Battle::Actions &, const Battle::Unit * );
+        virtual void HeroesAdd( const Heroes & );
+        virtual void HeroesRemove( const Heroes & );
+        virtual void HeroesPreBattle( HeroBase & );
+        virtual void HeroesAfterBattle( HeroBase & );
+        virtual void HeroesPostLoad( Heroes & );
+        virtual bool HeroesCanMove( const Heroes & hero );
+        virtual bool HeroesGetTask( Heroes & );
+        virtual void HeroesActionComplete( Heroes &, s32 );
+        virtual void HeroesActionNewPosition( Heroes & );
+        virtual void HeroesClearTask( const Heroes & );
+        virtual void HeroesLevelUp( Heroes & );
+        virtual bool HeroesSkipFog( void );
+        virtual std::string HeroesString( const Heroes & );
 
-    void HeroesAdd( const Heroes & );
-    void HeroesRemove( const Heroes & );
-    void HeroesTurn( Heroes & );
-    void HeroesMove( Heroes & );
-    bool HeroesGetTask( Heroes & );
-    void HeroesPreBattle( HeroBase & );
-    void HeroesAfterBattle( HeroBase & );
-    void HeroesPostLoad( Heroes & );
-    bool HeroesValidObject( const Heroes &, s32 );
-    bool HeroesCanMove( const Heroes & );
-    void HeroesAction( Heroes &, s32 );
-    void HeroesActionComplete( Heroes &, s32 );
-    void HeroesActionNewPosition( Heroes & );
-    void HeroesLevelUp( Heroes & );
-    void HeroesClearTask( const Heroes & );
-    bool HeroesSkipFog( void );
-    std::string HeroesString( const Heroes & );
+        virtual void CastleAdd( const Castle & );
+        virtual void CastleRemove( const Castle & );
+        virtual void CastlePreBattle( Castle & );
+        virtual void CastleAfterBattle( Castle &, bool attacker_wins );
 
-    void CastleAdd( const Castle & );
-    void CastleRemove( const Castle & );
-    void CastleTurn( Castle & );
-    void CastlePreBattle( Castle & );
-    void CastleAfterBattle( Castle &, bool attacker_wins );
+        virtual const char * Type( void ) const;
+        virtual const char * License( void ) const;
+        virtual void Reset();
+    };
 
-    const char * Type( void );
-    const char * License( void );
+    Base & Get( AI_TYPE type = SIMPLE );
+
+    void HeroesAction( Heroes & hero, s32 dst_index );
+    bool HeroesValidObject( const Heroes & hero, s32 index );
+    void HeroesMove( Heroes & hero );
 }
 
 #endif
