+++ src/fheroes2/battle/battle_main.cpp
@@ -53,7 +53,7 @@ Battle::Result Battle::Loader( Army & army1, Army & army2, s32 mapsindex )
         if ( army1.GetCommander()->isCaptain() )
             army1.GetCommander()->ActionPreBattle();
         else if ( army1.isControlAI() )
-            AI::HeroesPreBattle( *army1.GetCommander() );
+            AI::Get().HeroesPreBattle( *army1.GetCommander() );
         else
             army1.GetCommander()->ActionPreBattle();
     }
@@ -63,7 +63,7 @@ Battle::Result Battle::Loader( Army & army1, Army & army2, s32 mapsindex )
         if ( army2.GetCommander()->isCaptain() )
             army2.GetCommander()->ActionPreBattle();
         else if ( army2.isControlAI() )
-            AI::HeroesPreBattle( *army2.GetCommander() );
+            AI::Get().HeroesPreBattle( *army2.GetCommander() );
         else
             army2.GetCommander()->ActionPreBattle();
     }
@@ -106,7 +106,7 @@ Battle::Result Battle::Loader( Army & army1, Army & army2, s32 mapsindex )
     // after battle army1
     if ( army1.GetCommander() ) {
         if ( army1.isControlAI() )
-            AI::HeroesAfterBattle( *army1.GetCommander() );
+            AI::Get().HeroesAfterBattle( *army1.GetCommander() );
         else
             army1.GetCommander()->ActionAfterBattle();
     }
@@ -114,7 +114,7 @@ Battle::Result Battle::Loader( Army & army1, Army & army2, s32 mapsindex )
     // after battle army2
     if ( army2.GetCommander() ) {
         if ( army2.isControlAI() )
-            AI::HeroesAfterBattle( *army2.GetCommander() );
+            AI::Get().HeroesAfterBattle( *army2.GetCommander() );
         else
             army2.GetCommander()->ActionAfterBattle();
     }
